drf_api/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
djangorestframework==3.13.1
engineering-notation==0.6.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fpdf==1.7.2
gitdb==4.0.9
GitPython==3.1.27
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.9.2
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mock==4.0.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.4
notebook==6.4.11
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==22.1.1
plotly==5.7.0
plotly-express==0.4.1
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.28
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
pyrsistent==0.18.1
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
qrcode==7.3.1
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
setuptools==56.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.8.1
tenacity==8.0.1
terminado==0.13.3
tinycss2==1.1.1
tk-tools==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xlrd==2.0.1
xlwings==0.27.6
zipp==3.8.0
